+++ CHANGELOG.md
@@ -14,6 +14,13 @@ All notable changes to this RDK Service will be documented in this file.
 
 * Changes in CHANGELOG should be updated when commits are added to the main or release branches. There should be one CHANGELOG entry per JIRA Ticket. This is not enforced on sprint branches since there could be multiple changes for the same JIRA ticket during development. 
 
+## [3.4.0] - 2026-07-17
+### Fixed
+- Fixed the logging failure within the inprogress plugins
+- Added mutex before accessing the scan filters
+- Added t2 event for IPv6 Public IP
+- Added thread to publish the events without blocking the caller notification thread
+
 ## [3.3.0] - 2026-06-12
 ### Fixed
 - Fixed the memory leak that caused by the persistent m_nmClient and m_nmContext.